Replace all create task logic to interactor
